Cambio de directorio de trabajo y condicional para creación de usuario.

+++ src/107.1_01.d.Crear-registro-usuario.sh
@@ -12,13 +12,19 @@
 ## __Fuente__
 ###### [fuente]:(https://learning.lpi.org/es/learning-materials/102-500/107/107.1/107.1_01/)
 
-## __Importación de colores__
-source "${0%/*}"/000.Colores.sh
-
 ## __Configuración de variables__
 FQDN=$(hostname -f)
 USER='registro'
 timestamp=$(date +%F_%H.%M.%S)
+WD='/var/local/ubuntu-noble-server/src'
+
+## __Cambio de directorio de trabajo__
+###### Es necesario cuando el script es ejecutado a través de un enlace
+###### simbólico en cron.
+cd $WD
+
+## __Importación de colores__
+source $WD/000.Colores.sh
 
 ## __Respaldo de configuración__
 echo -e "$cian Respaldando configuración $default"
@@ -28,31 +34,34 @@ cp $DIR$FILE /var/local/backups/$FILE.$timestamp
 
 ## __Modificación de configuración__
 echo -e "$cian Modificando configuración $default"
-grep ferorge $DIR$FILE
+grep ferorge $DIR$FILE > /dev/null
 if [[ $? != 0 ]];then
 echo "
 ########################
 # Editado por ~ferorge #
 ########################
-"${0%/*}"/107.1_01.b.Creacion-usuario-pendiente.sh" >> $DIR$FILE
+$WD/107.1_01.b.Creacion-usuario-pendiente.sh" >> $DIR$FILE
 fi
 
-## __Crea una carpeta temporal__
-TMP_DIR=/tmp/temp
-mkdir $TMP_DIR
-
 ## __Crea usuario para crear usuarios mediante ssh__
-useradd --create-home -k $TMP_DIR --comment 'Registro de nuevos usuarios' --shell /home/$USER/registro.sh $USER
+id $USER > /dev/null
+if [[ $? != 0 ]];then
+  ## __Crea una carpeta temporal__
+  TMP_DIR=/tmp/temp
+  mkdir $TMP_DIR
+  ## __Crea el usuario__
+  useradd --create-home -k $TMP_DIR --comment 'Registro de nuevos usuarios' --shell /home/$USER/registro.sh $USER
+  ## __Establece la contraseña 'Ahora!'__
+  echo $USER:'Ahora!' | chpasswd
+  ## __Elimina la carpeta temporal__
+  rmdir $TMP_DIR
+fi
 
 ## __Copia el shell del usuario.__
-cp "${0%/*}"/107.1_01.a.Registro.sh /home/$USER/registro.sh
+cp $WD/107.1_01.a.Registro.sh /home/$USER/registro.sh
 chown registro:registro /home/$USER/registro.sh
 
-## __Establece la contraseña 'Ahora!'__
-echo $USER:'Ahora!' | chpasswd
-
 ## __Genera el cron para que elimine usuarios no confirmados.__
-ln -s ""${0%/*}"/107.1_01.c.Eliminacion-usuario-pendiente.sh" /etc/cron.daily/Eliminar-usuario-pendiente
+unlink /etc/cron.daily/Eliminar-usuario-pendiente
+ln -s "$WD/107.1_01.c.Eliminacion-usuario-pendiente.sh" /etc/cron.daily/Eliminar-usuario-pendiente
 
-## __Elimina la carpeta temporal__
-rmdir $TMP_DIR
